Refactor Uniform::[vp]sRegisterIndex to unsigned integers.
TRAC #22858

Author: Jamie Madill

git-svn-id: https://angleproject.googlecode.com/svn/branches/es3proto@2305 736b8ea6-26fd-11df-bfd4-992fa37f6226

+++ src/compiler/OutputHLSL.cpp
@@ -3100,7 +3100,7 @@ void OutputHLSL::declareUniform(const TType &type, const TString &name, int inde
 
     if (!structure)
     {
-        mActiveUniforms.push_back(Uniform(glVariableType(type), glVariablePrecision(type), name.c_str(), type.getArraySize(), index));
+        mActiveUniforms.push_back(Uniform(glVariableType(type), glVariablePrecision(type), name.c_str(), (unsigned int)type.getArraySize(), (unsigned int)index));
     }
     else
     {
